Refactored GT howtos
