--- doc/rtd/content/append1/bf/elem/howto.bf.004.rst
@@ -1,64 +1,0 @@
-.. _Howto BF 4:
-`Howto 04 - (Data) Store, plot, and save variables <https://github.com/fhswf/MLPro/blob/main/examples/bf/Howto%2004%20-%20(Data)%20Store%2C%20plot%2C%20and%20save%20variables.py>`_
-================
-Ver. 1.2.3 (2021-10-26)
-
-This module demonstrates how to store, plot, save and load variables. In the result section below, an example output shows the plots generated with the help of DataPlotting functionality of MLPro. With the help of Data module in MLPro, data can be stored as python dictionary objects, local binary files and csv files. The plots can also be saved as pdf files. The files can be saved and loaded from the path set.
-
-For this example implementation, following files will be saved in the path provided:
-    - model_loss.pdf
-    - reward.pdf
-    - states_1.pdf
-    - states_2.pdf
-    - data_storage.csv
-    - data_memory.pkl
-    - plot_memory.pkl
-
-The results also show a comparison between the data loaded from the csv and binary format.
-
-Results
-`````````````````
-
-The following figures show the plots generated by implementing the Data module in MLPro:
-
-.. image:: images/model_loss.png
-	:width: 49%
-.. image:: images/reward.png
-	:width: 49%
-.. image:: images/states_1.png
-	:width: 49%
-.. image:: images/states_2.png
-	:width: 49%
-
-This is the comparison between the data loaded from the binary and csv files using the load methods:
-
-::
-
-	Comparison :
-	Original data                : 2.73498
-	Loaded data from binary file : 2.73498
-	Comparison :
-	Original data             : 2.73498
-	Loaded data from csv file : 2.73498
-
-
-The above results can be generated by executing the script given in the example code section of this document.
-
-Prerequisites
-`````````````````
-
-Please install the following packages to run this examples properly:
-    - :ref:`MLPro <Installation>`
-    - `Matplotlib <https://pypi.org/project/matplotlib/>`_
-    
-
-Example Code
-`````````````````
-
-Run this code to generate the results as shown above
-
-**Note: For this example implementation, the values are generated by random, results may differ accordingly.**
-
-.. literalinclude:: ../../../../../../examples/bf/Howto 04 - (Data) Store, plot, and save variables.py
-    :language: python
-
