--- doc/rtd/content/append1/bf/elem/obs_howto.bf.001.rst
@@ -1,66 +1,0 @@
-.. _Howto BF 1:
-`Howto 01 - (Various) Logging (Obsolete) <https://github.com/fhswf/MLPro/blob/main/examples/bf/Howto-BF-002_(Various)_Timer.py>`_
-================
-Ver. 1.1.1 (2021-11-13)
-
-This is an example of implementing the Log class functionality of MLPro. The logs can be categorized into four different categories and can be logged at four different levels. A snippet of the output below shows logging information based on the log level set while initialzing the class. The following result can be generated by executing the python script in the example code section of this documents.
-
-Results
------------------
-This is an output to illustrate the Log class functionality in MLPro
-::
-		
-	--
-	-- Example 1: By default everything is logged...
-	--
-	
-	YYYY-MM-DD  HH:MM:SS:ssss  I  Demo class MyClass: Instantiated 
-	YYYY-MM-DD  HH:MM:SS:ssss  I  Demo class MyClass: Let me tell you what's going on... 
-	YYYY-MM-DD  HH:MM:SS:ssss  W  Demo class MyClass: Something is weird... 
-	YYYY-MM-DD  HH:MM:SS:ssss  E  Demo class MyClass: And here something failed... 
-	YYYY-MM-DD  HH:MM:SS:ssss  I  Demo class MyClass: But don't worry. Everything is fine. It's just a demo:) 
-	YYYY-MM-DD  HH:MM:SS:ssss  S  Demo class MyClass: This method terminated successfully!
-	 
-	
-	--
-	-- Example 2: Now logging is switched off...
-	--
-	
-	
-	--
-	-- Example 3: Only warnings and errors are logged...
-	--
-	
-	YYYY-MM-DD  HH:MM:SS:ssss  W  Demo class MyClass: Something is weird... 
-	YYYY-MM-DD  HH:MM:SS:ssss  E  Demo class MyClass: And here something failed... 
-	
-	--
-	-- Example 4: Only errors are logged...
-	--
-	
-	YYYY-MM-DD  HH:MM:SS:ssss  E  Demo class MyClass: And here something failed... 
-	
-	--
-	-- Example 5: Everything is logged again...
-	--
-	
-	YYYY-MM-DD  HH:MM:SS:ssss  I  Demo class MyClass: Let me tell you what's going on... 
-	YYYY-MM-DD  HH:MM:SS:ssss  W  Demo class MyClass: Something is weird... 
-	YYYY-MM-DD  HH:MM:SS:ssss  E  Demo class MyClass: And here something failed... 
-	YYYY-MM-DD  HH:MM:SS:ssss  I  Demo class MyClass: But don't worry. Everything is fine. It's just a demo:) 
-	YYYY-MM-DD  HH:MM:SS:ssss  S  Demo class MyClass: This method terminated successfully!
-
-Prerequisites
------------------
-
-Please install the following packages to run this examples properly:
-    - :ref:`MLPro <Installation>`
-
-Example Code
------------------
-
-Run this script to obtain the results as shown in the snippet above
-
-.. literalinclude:: ../../../../../../examples/bf/Howto-BF-002_(Various)_Timer.py.py
-    :language: python
-
