+++ doc/rtd/content/append1/howto.gt.rst
@@ -1,10 +1,10 @@
 Game Theory
-================
+===========
 
 The following examples demonstrate various functionalities of MLPro-GT:
 
 .. toctree::
    :maxdepth: 1
+   :glob:
    
-   gt/howto.gt.006
-   gt/howto.gt.007
+   gt/howto.*
