+++ src/mlpro/gt/examples/howto_gt_002_train_own_multi_player_with_multicartpole_game_board.py
@@ -1,24 +1,28 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : Howto-GT-001_Run_multi-player_with_own_policy_in_multicartpole_game_board.py
+## -- Module  : howto_gt_002_train_own_multi_player_with_multicartpole_game_board.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2021-06-06  0.0.0     DA       Creation
 ## -- 2021-06-06  1.0.0     DA       Release of first version
-## -- 2021-08-28  1.0.1     DA       Adjustments after changings on rl models
-## -- 2021-09-11  1.0.1     MRD      Change Header information to match our new library name
-## -- 2021-10-06  1.0.2     DA       Adjustments after changings on rl models
-## -- 2021-11-15  1.1.0     DA       Refactoring 
-## -- 2022-02-25  1.1.1     SY       Refactoring due to auto generated ID in class Dimension
+## -- 2021-07-01  1.1.0     DA       Extended by data logging/storing (user home directory)
+## -- 2021-07-06  1.1.1     SY       Bugfix due to method Training.save_data() update
+## -- 2021-08-28  1.1.2     DA       Adjustments after changings on rl models
+## -- 2021-09-11  1.1.2     MRD      Change Header information to match our new library name
+## -- 2021-09-28  1.1.3     SY       Adjustment due to implementation of SAR Buffer on player
+## -- 2021-10-06  1.1.4     DA       Refactoring 
+## -- 2021-11-16  1.2.0     DA       Refactoring 
+## -- 2021-12-07  1.2.1     DA       Refactoring 
+## -- 2022-02-25  1.2.2     SY       Refactoring due to auto generated ID in class Dimension
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.1.1 (2022-02-25)
-
-This module shows how to run an own multi-player with the enhanced multi-action game board 
-MultiCartPole based on the OpenAI Gym CartPole environment.
+Ver. 1.2.2 (2021-02-25)
+ 
+This module shows how to train an own multi-player with the enhanced multi-action
+game board MultiCartPole based on the OpenAI Gym CartPole environment.
 """
 
 
@@ -27,19 +31,18 @@ from mlpro.gt.models import *
 from mlpro.gt.pool.boards.multicartpole import MultiCartPolePGT
 import random
 import numpy as np
+from pathlib import Path
+import os
+from datetime import datetime
 
 
 
 
 
 # 1 Implement your own agent policy
-class MyPolicy (Policy):
+class MyPolicy(Policy):
 
     C_NAME      = 'MyPolicy'
-
-    def set_random_seed(self, p_seed=None):
-        random.seed(p_seed)
-
 
     def compute_action(self, p_state: State) -> Action:
         # 1 Create a numpy array for your action values 
@@ -55,7 +58,7 @@ class MyPolicy (Policy):
 
     def _adapt(self, *p_args) -> bool:
         # 1 Adapting the internal policy is up to you...
-        self.log(self.C_LOG_TYPE_W, 'Sorry, I am a stupid agent...')
+        self.log(self.C_LOG_TYPE_I, 'Sorry, I am a stupid agent...')
 
         # 2 Only return True if something has been adapted...
         return False
@@ -64,7 +67,7 @@ class MyPolicy (Policy):
 
 
 # 2 Implement your own game
-class MyGame (Game):
+class MyGame(Game):
 
     C_NAME      = 'Matrix'
 
@@ -79,7 +82,7 @@ class MyGame (Game):
         # 2.1 Create empty Multi-Player
         multi_player = MultiPlayer(
             p_name='Human Beings',
-            p_ada=True,
+            p_ada=p_ada,
             p_logging=p_logging
         )
 
@@ -91,36 +94,39 @@ class MyGame (Game):
                 p_policy=MyPolicy(
                     p_observation_space=self._env.get_state_space().spawn([ss_ids[0],ss_ids[1],ss_ids[2],ss_ids[3]]),
                     p_action_space=self._env.get_action_space().spawn([as_ids[0]]),
-                    p_ada=True,
+                    p_buffer_size=1,
+                    p_ada=p_ada,
                     p_logging=p_logging
                 ),
                 p_name='Neo',
                 p_id=0,
-                p_ada=True,
+                p_ada=p_ada,
                 p_logging=p_logging
             ),
             p_weight=0.3
         )
 
 
-        # 2.3 Add Single-Player #2 with own policy (controlling sub-environments #2,#3)
+        # 2.2 Add Single-Player #2 with own policy (controlling sub-environments #2,#3)
         multi_player.add_player(
             p_player=Player(
                 p_policy=MyPolicy(
                     p_observation_space=self._env.get_state_space().spawn([ss_ids[4],ss_ids[5],ss_ids[6],ss_ids[7],ss_ids[8],ss_ids[9],ss_ids[10],ss_ids[11]]),
                     p_action_space=self._env.get_action_space().spawn([as_ids[1],as_ids[2]]),
-                    p_ada=True,
+                    p_buffer_size=1,
+                    p_ada=p_ada,
                     p_logging=p_logging
                 ),
                 p_name='Trinity',
                 p_id=1,
-                p_ada=True,
+                p_ada=p_ada,
                 p_logging=p_logging
             ),
             p_weight=0.7
         )
 
-        # 2.4 Adaptive ML model (here: our multi-player) is returned
+
+        # 2.3 Return multi-player as adaptive model
         return multi_player
 
 
@@ -130,20 +136,25 @@ class MyGame (Game):
 
 if __name__ == "__main__":
     # 3.1 Parameters for demo mode
-    logging     = Log.C_LOG_ALL
+    cycle_limit = 200
+    logging     = Log.C_LOG_WE
     visualize   = True
-  
+    path        = str(Path.home())
+ 
 else:
     # 3.2 Parameters for internal unit test
+    cycle_limit = 10
     logging     = Log.C_LOG_NOTHING
     visualize   = False
+    path        = None
 
 
-mygame  = MyGame(
-    p_mode=Mode.C_MODE_SIM,
-    p_ada=True,
-    p_cycle_limit=200,
-    p_visualize=visualize,
-    p_logging=logging )
+# 3.3 Create and run training object
+training = GTTraining(
+        p_game_cls=MyGame,
+        p_cycle_limit=cycle_limit,
+        p_path=path,
+        p_visualize=visualize,
+        p_logging=logging )
 
-mygame.run()
+training.run()
