+++ src/mlpro/rl/examples/Howto-RL-014_Advanced_training_with_stagnation_detection.py
